requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
pytest==6.2.5
pytest-cov==3.0.0
scipy==1.7.3
numpy==1.21.6
pydantic==1.10.7